requirements.txt
streamlit
psycopg2-binary
sqlglot
pyyaml
langchain
sentence-transformers
faiss-cpu
numpy
pillow
python-docx
pypandoc
pymupdf
transformers
